requirements.txt
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
beanie==1.26.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
coverage==7.6.1
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.19.0
factory_boy==3.3.1
Faker==30.0.0
fastapi==0.115.0
fastapi-cache2==0.2.2
flake8==7.1.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lazy-model==0.2.0
limits==3.13.0
locust==2.31.8
MarkupSafe==2.1.5
mccabe==0.7.0
motor==3.6.0
msgpack==1.1.0
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
pluggy==1.5.0
psutil==6.0.0
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pymongo==4.9.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.10
pyzmq==26.2.0
redis==5.1.0
requests==2.32.3
rsa==4.9
setuptools==75.1.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.38.6
time-machine==2.15.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zope.event==5.0
zope.interface==7.0.3
